qualcommax: nbg7815: update fan support to be enable by default
An update is done to enable fan by default and on every reboot. With
change made for kernel 6.6 fan was disable after a clean upgrade. This
device can be very hot if the 3 radios are enable and 10g are in use, so
for safety, fan has been enabled by default.